# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sensitive Stock Backend API",
    "version": "0.1.0",
    "description": "Global OpenAPI contract for the Sensitive Stock Flask backend. All protected business operations use the shared bearerAuth scheme."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/market": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketOverview",
        "summary": "Return market source metadata and available market routes.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketOverviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/quotes": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketQuotes",
        "summary": "Return latest quotes for one or more A-share symbols.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketQuotesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "symbols",
            "in": "query",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "minItems": 1,
              "title": "Symbols",
              "type": "array"
            },
            "style": "form",
            "explode": false
          }
        ]
      }
    },
    "/api/v1/market/sectors": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketSectors",
        "summary": "Return hot concept or industry sectors.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketSectorsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "name": "sectorType",
            "in": "query",
            "required": false,
            "schema": {
              "default": "concept",
              "enum": [
                "concept",
                "industry"
              ],
              "title": "Sectortype",
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/market/news": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketNews",
        "summary": "Return latest Jin10 market news items.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketNewsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Symbols",
              "type": "array"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "thinking",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "enabled",
                    "disabled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Thinking"
            }
          },
          {
            "name": "reasoningEffort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "high",
                    "max"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Reasoningeffort"
            }
          }
        ]
      }
    },
    "/api/v1/market/news/intelligence": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketNewsIntelligence",
        "summary": "Return derived market intelligence from Jin10 news and market data.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketNewsIntelligenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Symbols",
              "type": "array"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "thinking",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "enabled",
                    "disabled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Thinking"
            }
          },
          {
            "name": "reasoningEffort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "high",
                    "max"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Reasoningeffort"
            }
          }
        ]
      }
    },
    "/api/v1/market/news/predictions": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketNewsPredictions",
        "summary": "Return multi-source market news predictions and backtest handoff.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketNewsPredictionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "name": "symbols",
            "in": "query",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Symbols",
              "type": "array"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "thinking",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "enabled",
                    "disabled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Thinking"
            }
          },
          {
            "name": "reasoningEffort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "high",
                    "max"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Reasoningeffort"
            }
          }
        ]
      }
    },
    "/api/v1/market/news/prediction-history": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketPredictionHistory",
        "summary": "Return recent stored market prediction runs.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/v1/market/news/predictions/{runId}": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "getMarketPredictionDetail",
        "summary": "Return one stored market prediction run.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/market/news/predictions/{runId}/evaluate": {
      "get": {
        "tags": [
          "market"
        ],
        "operationId": "evaluateMarketPredictionRun",
        "summary": "Evaluate one stored market prediction run against latest quotes.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionEvaluationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected backend error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token returned by POST /api/v1/auth/login."
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "additionalProperties": false
      },
...
      "MarketOverviewResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceMetadata"
          },
          {
            "type": "object",
            "required": [
              "primarySource",
              "fallbackSources",
              "routes"
            ],
            "properties": {
              "routes": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "MarketQuotesResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceMetadata"
          },
          {
            "type": "object",
            "required": [
              "source",
              "items"
            ],
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MarketQuote"
                }
              }
            }
          }
        ]
      },
...
      "MarketSectorsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceMetadata"
          },
          {
            "type": "object",
            "required": [
              "source",
              "sectorType",
              "items"
            ],
            "properties": {
              "sectorType": {
                "type": "string",
                "enum": [
                  "concept",
                  "industry"
                ]
              },
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MarketSector"
                }
              }
            }
          }
        ]
      },
...
      "MarketNewsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceMetadata"
          },
          {
            "type": "object",
            "required": [
              "source",
              "requestedLimit",
              "degraded",
              "items"
            ],
            "properties": {
              "sourceQuality": {
                "$ref": "#/components/schemas/MarketNewsSourceQuality"
              },
              "dedupeMetadata": {
                "$ref": "#/components/schemas/MarketNewsDedupeMetadata"
              },
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MarketNewsItem"
                }
              }
            }
          }
        ]
      },
...
      "MarketNewsIntelligenceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MarketNewsResponse"
          },
          {
            "type": "object",
            "required": [
              "keywords",
              "sectorHints"
            ],
            "properties": {
              "keywords": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              },
              "sectorHints": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        ]
      },
...
      "MarketNewsPredictionsResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MarketNewsIntelligenceResponse"
          },
          {
            "type": "object",
            "required": [
              "runId",
              "createdAt",
              "channels",
              "predictionMetadata",
              "predictions",
              "riskNotes",
              "backtestHandoff"
            ],
            "properties": {
              "runId": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              },
              "channels": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MarketNewsChannel"
                }
              },
              "sourceQuality": {
                "$ref": "#/components/schemas/MarketNewsSourceQuality"
              },
              "dedupeMetadata": {
                "$ref": "#/components/schemas/MarketNewsDedupeMetadata"
              },
              "predictionMetadata": {
                "$ref": "#/components/schemas/MarketPredictionMetadata"
              },
              "predictions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MarketPrediction"
                }
              },
              "predictionSummary": {
                "type": "string"
              },
              "riskNotes": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "backtestHandoff": {
                "$ref": "#/components/schemas/BacktestHandoff"
              }
            }
          }
        ]
      },
      "PredictionHistoryResponse": {
        "type": "object",
        "required": [
          "items",
          "metadata"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PredictionHistoryRun"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/CapabilityMetadata"
          }
        },
        "additionalProperties": true
      },
...
      "PredictionDetailResponse": {
        "type": "object",
        "required": [
          "runId",
          "createdAt",
          "predictions",
          "predictionMetadata"
        ],
        "properties": {
          "runId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketNewsItem"
            }
          },
          "sourceQuality": {
            "$ref": "#/components/schemas/MarketNewsSourceQuality"
          },
          "dedupeMetadata": {
            "$ref": "#/components/schemas/MarketNewsDedupeMetadata"
          },
          "predictionMetadata": {
            "$ref": "#/components/schemas/MarketPredictionMetadata"
          },
          "predictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketPrediction"
            }
          },
          "riskNotes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "backtestHandoff": {
            "$ref": "#/components/schemas/BacktestHandoff"
          }
        },
        "additionalProperties": true
      },
      "PredictionEvaluationResponse": {
        "type": "object",
        "required": [
          "runId",
          "evaluatedAt",
          "evaluationSummary",
          "evaluationItems",
          "metadata"
        ],
        "properties": {
          "runId": {
            "type": "string"
          },
          "evaluatedAt": {
            "type": "string"
          },
          "evaluationSummary": {
            "$ref": "#/components/schemas/PredictionEvaluationSummary"
          },
          "evaluationItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PredictionEvaluationItem"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/CapabilityMetadata"
          }
        },
        "additionalProperties": true
      },
...
    },
...
  },
...
}